% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IMD.r
\name{getIMD}
\alias{getIMD}
\title{get Index of Multiple Deprivation data}
\usage{
getIMD(path = .defaultPath, file = .defaultFile, scores.only = TRUE)
}
\arguments{
\item{path}{path to file}

\item{file}{filename of the .csv file containing everything}

\item{scores.only}{TRUE (default) to strip ranking and deciles, FALSE to keep everything}
}
\value{
IMD dataset
}
\description{
load and prepare the data from .csv file
}
\examples{
IMD2015 <- getIMD()
}

