\name{imd-package}
\alias{imd-package}
\alias{imd}
\docType{package}
\title{Index of Multiple Deprivation}
\description{
Stuff to make life easier with IMD data
}
\author{
Author: Wei Liang Yap

Maintainer: Wei Liang Yap <yapweiliang@yahoo.co.uk>
}
\references{
https://www.gov.uk/government/statistics/english-indices-of-deprivation-2015
}
\keyword{ package }
\seealso{
}
\examples{
IMD2015 <- getIMD()
}

